winwave: fix conversion warnings
VS2019:

3>play.c
3>.\baresip\modules\winwave\play.c(174,45): warning C4267: '=': conversion from 'size_t' to 'WORD', possible loss of data
3>src.c
3>.\baresip\modules\winwave\src.c(163,45): warning C4267: '=': conversion from 'size_t' to 'WORD', possible loss of data

+++ modules/winwave/play.c
@@ -171,7 +171,7 @@ static int write_stream_open(struct auplay_st *st,
 	wfmt.nChannels       = prm->ch;
 	wfmt.nSamplesPerSec  = prm->srate;
 	wfmt.wBitsPerSample  = (WORD)(st->sampsz * 8);
-	wfmt.nBlockAlign     = prm->ch * st->sampsz;
+	wfmt.nBlockAlign     = (WORD)(prm->ch * st->sampsz);
 	wfmt.nAvgBytesPerSec = wfmt.nSamplesPerSec * wfmt.nBlockAlign;
 	wfmt.cbSize          = 0;
